move db into util
